+++ CHANGELOG.md
@@ -2,5 +2,4 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Percursus/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Percursus/releases)
 
-- Fixed: Access to secret values when opening the race overview is prevented [retail]
-- Minor code adjustments [retail]
+- Added: New Race Tracker background 'Midnight' [retail]